+++ lib/urllib.js
@@ -10,7 +10,6 @@ var zlib = require('zlib');
 var ua = require('default-user-agent');
 var digestAuthHeader = require('digest-header');
 var util = require('util');
-var EventEmitter = require('events').EventEmitter;
 var ms = require('humanize-ms');
 var statuses = require('statuses');
 var contentTypeParser = require('content-type');
@@ -889,102 +888,6 @@ function parseJSON(data, fixJSONCtlChars) {
   return result;
 }
 
-
-function HttpClient(options) {
-  EventEmitter.call(this);
-  options = options || {};
-
-  if (options.agent !== undefined) {
-    this.agent = options.agent;
-    this.hasCustomAgent = true;
-  } else {
-    this.agent = exports.agent;
-    this.hasCustomAgent = false;
-  }
-
-  if (options.httpsAgent !== undefined) {
-    this.httpsAgent = options.httpsAgent;
-    this.hasCustomHttpsAgent = true;
-  } else {
-    this.httpsAgent = exports.httpsAgent;
-    this.hasCustomHttpsAgent = false;
-  }
-}
-util.inherits(HttpClient, EventEmitter);
-
-HttpClient.prototype.request = HttpClient.prototype.curl = function (url, args, callback) {
-  if (typeof args === 'function') {
-    callback = args;
-    args = null;
-  }
-  args = args || {};
-  args.emitter = this;
-  args.agent = getAgent(args.agent, this.agent);
-  args.httpsAgent = getAgent(args.httpsAgent, this.httpsAgent);
-  return exports.request(url, args, callback);
-};
-
-HttpClient.prototype.requestThunk = function (url, args) {
-  args = args || {};
-  args.emitter = this;
-  args.agent = getAgent(args.agent, this.agent);
-  args.httpsAgent = getAgent(args.httpsAgent, this.httpsAgent);
-  return exports.requestThunk(url, args);
-};
-
-
-exports.HttpClient = HttpClient;
-
-
-function HttpClient2() {
-  HttpClient.call(this);
-}
-
-util.inherits(HttpClient2, HttpClient);
-
-HttpClient2.prototype.request = HttpClient2.prototype.curl = function request(url, args) {
-  var self = this;
-  args = args || {};
-  args.retry = args.retry || 0;
-  args.isRetry = args.isRetry || function(res) {
-    return res.status >= 500;
-  };
-  return HttpClient.prototype.request.call(self, url, args)
-  .catch(function(err) {
-    if (args.retry > 0) {
-      args.retry--;
-      debug('retry request %s, remain %s, err %s', url, args.retry, err);
-      return self.request(url, args);
-    }
-    throw err;
-  })
-  .then(function(res) {
-    if (args.retry > 0 && typeof args.isRetry === 'function' && args.isRetry(res)) {
-      args.retry--;
-      debug('retry request %s, remain %s', url, args.retry);
-      return self.request(url, args);
-    }
-    return res;
-  });
-};
-
-HttpClient2.prototype.requestThunk = function requestThunk(url, args) {
-  var self = this;
-  return function(done) {
-    self.request(url, args)
-    .catch(done)
-    .then(function(res) {
-      done(null, res);
-    });
-  };
-};
-
-exports.HttpClient2 = HttpClient2;
-
-
-exports.create = function (options) {
-  return new HttpClient(options);
-};
 
 /**
  * decode response body by parse `content-type`'s charset
